requirements.txt

asgiref==3.4.1
Django==3.2
django-admin-volt==0.0.5
django-ckeditor==6.2.0
django-js-asset==1.2.2
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5

